+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - added error code for checking update endorser(endorser has to be whitelisted)
 - added function for updating nft endorser list
 - update rpc endpoint for debug/fee_info to display nft and ft multiplier
+- fee with token action (nft/ft) will return default fee if not found
 
 # Version 1.2.1
 - Adding non fungible token module